pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]
